마이페이지 재작업을 위한 컴포넌트 제거
